Move GenericTypeParamBoundMismatch to resolver (#3087)
This is super weird. I didn't realize that resolver worked this way (it
dispatches a call manually from type_syntax parsing), which means that
there is at least one intrinsic in calls.cc that can report errors in
`typed: false` files. I decided to move the error code to resolver to
make that more clear.

Separately, I think we should fix this, so that we don't double report
those errors.