Parallelize namer (#2986)
* Parallelizes the namer typechecking pass.

Namer inserts symbols into the symbol table. It's currently completely serial, which is a scalability bottleneck on large projects.

This change breaks up Namer into three passes:

* SymbolFinder (parallel): AST traversal that locates names that need to be defined as symbols.
* SymbolDefiner (serial): Consumes a flat list of symbols to define from SymbolFinder, and defines them. Although serial, it should be significantly faster than the previous serial namer pass since it does not perform any AST traversal.
* TreeSymbolizer (parallel): Rewrites the AST to include references to the now-defined symbols.

In addition to improving performance, the output of SymbolFinder is cacheable; it contains flat vectors with structs that are trivial to serialize. SymbolFinder does not produce any errors, so typechecking starting from the flat vectors in SymbolDefiner should produce the same errors as starting from SymbolFinder.

* Bump up magic queue size constant.

We're doing more stuff in parallel now, so the queue can get backed up with lambdas to run. Still, it's not great that this can happen in the first place -- we should consider making this abstraction safer so that no threads are left behind.