Preserve the generated error class on serialize
We do some fancy wrapping logic that tries to provide more context on
where the error message came from when it came from the generated
serialize, probably because the line numbers look terrible with the
module_eval tricks we're doing.

The problem is that this means sometimes the error kind can change
depending on the order the code loads.

The change to the test exemplifies this. We swallowed a NoMethodError
and converted it to a TypeError.