deepCopy return type instead of moving
In the DefaultArgs rewriter, `overrides` are translated to voids.
However, this causes issues are there are both a `params` and
`override` (void) in the same signature.

When params are hit, the return type is moved into the send's args. When
the override is later hit, the return type is now nil (because of the
move), this means that an empty tree is returned from `mangleSig`. This
causes the sig for the override not to be written.

As the signature isn't written, this causes the defaultArg to return
`T.untyped`. Then later in interference, anything compared to untyped
will pass, hiding the error.