Resolve all instance variable warnings
Sorbet Runtime generates a lot of warnings like
`warning: instance variable @foo not initialized` due to lazy
initialization of various instance variables.

This commit attempts to fix most of these that can be identified by
running the test suite.

Coincidentally, this investigation also lead to finding and fixing a
memoization attempt using `||= false` which would be executed every
time and defeat the aim to memoize the ivar.