Remove Singleton / final! from T::Enum DSL
It's more convenient to just special case classes with T::Enum names in
`isSingleton` than it is to deal with Singleton.

(e.g., MySingleton.instance doesn't forward args to the parent, so we
had that hack where we left a call to the original `new` around in the
tree to be sure that it would be type checked)

By not using `include Singleton` and `final!`, we don't have to
compromise quite as much.

This also means that it's no longer the case that

    MyEnum::A.is_a?(Singleton)

which was an awkward relic of the old desugaring.