Get C level backtraces in working in vm_dump.c (#139)
* Get C level backtraces in working in vm_dump.c

This is a hack to get C level backtraces working. (The default autoconf
test for backtraces uses a sigaltstack size that is too small, so the
SIGSEGV signal handler itself causes a SIGSEGV). This value was plucked
from signal.c:

<https://github.com/ruby/ruby/blob/v2_6_5/signal.c#L568>

The RubyVM uses C macro defines to feature detect whether `backtrace(2)`
support is available, and if so it includes C level backtraces when the
RubyVM itself crashes.

The specific variables it looks at:

- `HAVE_BACKTRACE`

  this is a macro defined by autoconf because there is a line in the
  configure script like `AC_CHECK_FUNCS(backtrace)` (see the autoconf
  docs for more).

- `BROKEN_BACKTRACE`

  this comes from a custom program that Ruby's configure script runs to
  attempt to figure out whether actually using `backtrace(2)` in a real
  program works. You can see the autoconf program here.

  <https://github.com/ruby/ruby/blob/v2_6_5/configure.ac#L2817-L2863>

  It uses `sigaltstack` and `SA_ONSTACK` to create a seperate stack for
  handling signals.

The problem was: `SIGSTKSZ` (which comes from a system header!) was not
suggesting a large enough stack size. When checking on an Ubuntu 16.04
box, we found that `SIGSTKSZ` was 8192 and `MINSIGSTKSZ` was 2048.

* whoops