Add a test that crashes unless we pin the keywordArgsSingleton Hash (#165)
This test is designed to have prevented #159 from being merged, so that
we wouldn't have needed to revert it.

This test is pretty wild. We need two files, because the `Init_`
function has to return to manifest the bug. Since `Init_` is in charge
of creating the keywordArgsSingleton Hash in the first place, it has a
reference to it on its C stack for the duration of the call.

Once the `Init_` function is popped, that reference isn't there anymore,
and Ruby is free to garbage collect the keywordArgsSingleton (except of
course, we pin it now, to fix the bug).

That's why there has to be two files.