Add CreateGEP comment, if -> switch, sendArgArrays -> sendArgArrayByBlock (#166)
* Add CreateGEP comment, if -> switch, sendArgArrays -> sendArgArrayByBlock

The `switch` gets us exhaustiveness (the `if` was already exhaustive,
but I didn't know that by looking at it, whereas with the switch it's
guaranteed by clang).

I renamed the variable to make it easier to grep for. That's what the
field is called in IREmitterContext. We never use the name verbatim when
setting the field, because it's only set with a curly brace initializer.

There are only a handful of CreateGEP instructions in the whole program,
and this was the one I ran into first, so maybe that means it's the one
that other people are likely to run into first.

* Fix comments

* Update comment