Process Id returned from wait methods can be nil
Follow-up to: https://github.com/sorbet/sorbet/pull/3115

There is a rare but valid case where a certain combination of `pid`
values and `flags` to the `Process.wait` (and friends) will make a `nil`
return value possible.

The previous attempt at fixing the signature had only done it for
`Process.wait2` but it should be applied to `Process.wait`,
`Process.waitpid` and `Process.waitpid2` as well.