Replace most maps in CFG with vectors (#3346)
This PR replaces most maps/sets in CFG with vectors for a significant speedup on individual files with many basic blocks/variables (70%+ faster on one particular input I iterated on). This translates into a noticeable end-to-end speedup on large codebases and improves IDE responsiveness.

I was able to do this by introducing a LocalRef abstraction, which is a CFG-local reference to a local variable. Like other Ref classes, a LocalRef has a numerical ID that is the index into a vector of LocalVariables. A LocalRef is also half the size of a LocalVariable (4 bytes).

I use LocalRef for two data structures that were key to the speedup:

* Vectors of size # of locals that act like maps (vector[LocalRef.id] = value for that LocalRef). These can be preallocated and support insertion/deletion with no further memory allocation at the cost of slower iteration thru elements in the vector.
* Sorted vectors that contain LocalRef IDs, which act as sets. These are great for sparse sets that code needs to iterate through. It's cheap to take the union of two of these (linear time). These are less good for lookup (O(n) or O(logn) if using binary search).

While working on this change, I made a few additional common case optimizations.