Add CFG to IREmitterContext (#185)
After 0e1db4aa to introduce cfg::LocalRef's, there are a lot of call
sites that currently accept an IREmitterContext but that don't accept a
CFG. In particular, there's a deeply nested call site I want to change
to look up the name of a local variable, but I can't without threading a
CFG all the way down there.

For pretty much all of IREmitter, there's only one sensible CFG to have:
the CFG of the method that's currently being emitted. So instead of
storing a SymbolRef of the method, store the CFG (which also has a
SymbolRef inside it).