Allow vanilla ruby installations (#3257)
* Allow vanilla ruby installations

The previous logic did not find gems in a default ruby/rubygems
installation. For example the default path for user-installed gems on
macOS is `$HOME/.gem/ruby/2.6.0`. This was matched by the very first
pattern, `/^.*\/(ruby)\/([\d.]+)\//`, determining that this is the
location of a gem called `ruby` version 2.6.0, which is wrong.

This revision adds a much more general pattern which matches the default
gem path, and also customised BUNDLE_PATH (for users installing gems
into `$MY_PROJECT/bundle` for example). The new pattern must come before
the `# ruby` pattern for the aforementioned reason, subsuming the
previous `# gem` pattern; so that has been removed.

* Tighten the regex in accordance with the original

The new regex was undesriably accepting. Tightened to support only the
use cases intended.