Remove synthesizeExpr when making default args Cast
This change should have zero user-visible impact.

There are two places in Sorbet where we use `synthesizeExpr` to make a
Cast instruction, but there's only one place where `isSynthetic` on a
Cast instruction is important.

This is the call site where it isn't important as far as I can tell.
There are some changes I need to make to this code for a future feature,
and this is pre-work that makes that change simpler.