Skip typeTest for <self> (#194)
* Skip typeTest for <self>

It's impossible to assign into `<self>` (in Ruby `self` is a keyword so
`self = ...` is a syntax error).

This means that we can use whether `bind.bind.variable` is `<self>` to
be the judge of whether the `Cast` instruction we see is the initial
`<self> = Cast { var = <self> , ... }` instruction that we synthesize
at the top of all methods when building a CFG.

Also, this is a TypeError in Ruby:

    class A
      def foo; end
    end

    class B; end

    A.instance_method(:foo).bind(B.new).call

with the output:

    foo.rb:7:in `bind': bind argument must be an instance of A (TypeError)

So the Ruby VM will catch this anyways, even if we don't emit the
typeTest.

The case of self type tests within Proc's is different. Sorbet currently
uses the `LoadSelf` instruction to load the `<self>` variable inside a
block. This PR doesn't touch that codepath (the compiler might be broken
right now on weird self values inside blocks, but it's not made more
broken by this PR).

* Add a benchmark that just calls a compiled method

This method has no body and no sig and it's an instance method on an
instance stored in a local variable, so I think that this is the
smallest test case I can write that would benefit from no self type
checking, which should mean that this represents somewhat close to the
best case as far as impact from removing self type checks.

master:

    source                          interpreted  compiled
    stripe/while_10_000_000.rb      .205         .049
    stripe/call_compiled_method.rb  .355         .252

this branch:

    source                          interpreted  compiled
    stripe/while_10_000_000.rb      .205         .048
    stripe/call_compiled_method.rb  .347         .226

* Add a comment

* update testdata exp