Fix `hidden-methods` failure in Ruby 2.7 (#3422)
Resolves #2771

Due to the way argument forwarding is implemented in Ruby 2.7, at
runtime Sorbet will see parameter names for the expanded method
parameters that are invalid. This causes a failure in the hidden methods
generation process since the output ends up being an invalid RBI file.

The fix is to sanitize the parameter names to be valid unique names so
that we still capture the runtime method signature but also have valid
parameter names as well.