Allow method hooks to be disabled globally
Recently, we've had a case in our codebase where we experienced a 100%
slowdown in an endpoint due to the addition of a signature in one of our
Active Record models. My research into what caused the problem lead me to
realize that the method was triggering many Active Record callback
methods to be defined on the returned object for every call of the
method. This was causing all those methods to trigger the method added
hooks since Sorbet was activated on the model class, thus causing the
slowdown.

While I am still investigating the root cause of the problem in our
codebase (i.e. understanding why all those callback methods are being
added in the first place), I also wanted to make Sorbet behaviour more
deterministic in production by adding an optional method to disable
method hooks.

Afterall, in production, we eager load our application, and unwrap all
of our signatures after boot. Thus, we don't expect any methods with
signatures to be defined after that point, and can safely disable the
method hooks. The added `T::Utils.disable_method_hooks` method should
allow us to do that.