Deindent symbol based intrinsics (#210)
* untangle the intrinsic applicability check

Inverting the tests and rewording them slightly makes it clearer when we can
continue to considering the next intrinsic and avoids rightward code drift.

* invert the type check for checking class applicability

This change lets us deindent the core logic for setting up the (next) type
check for the intrinsic applicability.

* untangle the name-based intrinsic applicability check

This change makes the name-based intrinsics mirror the symbol-based
intrinsics.  It's debatable which style is preferred here, since the change
introduces more lines, but at least the intrinsic codepaths are using
similar styles of checks now.