Make VSCode symbol search work for nested ruby (#3475)
* Make VSCode symbol search work for nested ruby

VSCode has its own internal ranking of returned search results that
compares the search string against the returned symbol name (ignoring
the container). Part of this symbol search short-circuits if the query
string is not a subsequence of the result string. Therefore `Foo::Bar`
query will short-circuit on the result `Bar` since `Foo::` does not
appear in `Bar`. To work around this issue, return the full name as the
name to appease VSCode.

* Update tests

Co-authored-by: Kevin Miller <kmiller@stripe.com>