Allow extend T::Sig at the top level (#3491)
* Allow extend T::Sig at the top level

This is particularly handy for scripts, where you just want to write
everything at the top-level of a file.

Sorbet already type checks using `extend T::Sig` and `sig` at the top
level of a file correctly, so it's annoying to have it not work when I
copy/paste an example into a file and run it.

I very much don't want to encourage that people do this except in small
files. In large projects, if there are lots of top-level defined
methods, Sorbet will think they are all defined and available because it
has no way to know which files are the entrypoints to a program and will
thus never be run at the same time as other files.

But for small projects I think this is pretty nice. Especially one-off
scripts.

* Add a test