Allow to pass doctest specific options to each runner
Use the DOCTEST_CONFIG_NO_UNPREFIXED_OPTIONS[0] define to allow passing doctest specific options (which are prefixed with "--dt-" to avoid colliding with any current/future custom command line flags the runner may have).

As a rationale, this allows to use options such as "--no-colors" (specified as "--dt-no-colors") to force disable escape sequences when doctest TTY recognition fails to work (which is the case if you are trying to debug from VSCode).

[0] https://github.com/onqtam/doctest/blob/master/doc/markdown/configuration.md#doctest_config_no_unprefixed_options