Add minimized test case for bug that we accidentally fixed
The changes in #3505 accidentally fixed Sorbet's behavior on this file.
The old behavior did not report an error. I'm not sure why, but I think
it has something to do with the changes to `calls.cc` (changing
untyped -> noSymbol).

In any case, it's a nice edge case with union types and intersection
tyoes and generics so I figured it might come in handy in preventing
future regressions.