Add VSCode debug configuration to debug LSP tests
Add a sample VSCode debug launch.json to be able to debug LSP unit tests (can be easily duplicated for other runners)

The two key additions to the configuration from the default templates are the "setupCommands" and "sourceFileMap" options. The first one make sure that .lldbinit is properly sourced in the session (which will set a proper source map at the LLDB level). The second one will ensure that the VSCode *UI* correctly maps breakpoints/stacktrace information from LLDB to the workspace (why would it be not reading the LLDB source map setting and using that? no clue).