Use a more efficient set representation in CFG (#3526)
Use a more efficient set representation in CFG.

Previously, we used a vector of sorted integers and a vector of booleans in  different places. The former enables fast iteration, but slow (log n) lookup and insertion. The latter enables fast lookup, but slow iteration.

This PR introduces a `UIntSet` class that represents the set as a vector of u4s. Like a vector of booleans, membership is represented by a single bit, enabling fast lookup and compact representation. Like a vector of sorted integers, iteration is relatively fast as we can check entire integers for bits all at once (which is good for the common-case of an empty set). The size of the set can be computed quickly using the popcount CPU instruction.

There is a unit test suite that checks the correctness of `UIntSet`.

I placed `UIntSet` into `common` because I expect it to be generally useful. I am happy to move it into `cfg`, though, until it is used cross-package.

This change already improves Sorbet performance on Stripe's codebase. Some follow-on investigations (for a future PR) include:
* Using `UIntSet` for the type test reverse index in infer.