add moves for rvalue references in constructors (#3530)
Due to the wonders of C++, simply declaring rvalue reference arguments
in your constructor is not enough to be maximally efficient; one must
also `move` those arguments into the relevant places.  sorbet was not
doing this in all places.

The changes in LSPProcessor.cc are to fix the use-after-move errors
once we started taking advantage of the rvalue reference in `File::File`.