`initial` parameter of `inject` can be of any type
The current RBI for `Enumerable#inject` assumes that the `initial`
parameter for the overload with a block given will always be of type
`Enum`, but that is not always the case.

For example:

```ruby
[1,2,3].inject("") { |str, num| str << num.to_s }
```

is perfectly valid. The type of `str` should be `String` and the
`inject` call should return a `String`, not an `Integer`.