Add natvis file and plumbing to allow pretty-printing some constructs
This changeset adds support for visualizing Name and NameRefs in *some* circumstances.

For Name this will be the case when the underlying storage is a direct utf8 buffer, for NameRef the value will be displayed if it's pointing to a well-known entry (for which generate_names now produces a debug lookup table that's only included in debug mode).