remove ENFORCEs for Ruby stringification and interning  (#227)
* remove ENFORCEs for Ruby stringification and interning

Removing `ENFORCE`s is not something done lightly, so here's the broader
context:

For literal hashes (and arrays, and maybe a few other things), we currently
build the hash (resp. array) each time the hash is requested.  The Ruby VM's
bytecode, however, builds the literal hash once and then simply duplicates
it each time the hash is requested.  This duplication is relatively faster
than building because you don't have to recompute hashes for all the keys,
etc.

The way you'd implement this is to have a separate function that looks
something like this:

```
// We are constructing {"a": "b", "c": "d"}
VALUE globalHashLiteral;

void buildGlobalHashLiteral() {
  ...emit code to reference the VALUE for "a"
  ...emit code to reference the VALUE for "b"
  ...emit code to reference the VALUE for "c"
  ...emit code to reference the VALUE for "d"
  ...store all those references in a stack-allocated array
  globalHashLiteral = sorbet_hashBuild(4, array);
}
```

and then whenever you want to reference the hash literal, you can just call
`sorbet_hashDup(globalHashLiteral)`.  (You can implement this with the
moral equivalent of function-local statics, but the implementation at the
LLVM IR level is basically the above.)

We want to reuse the code in Payload.cc for building global variables for
`"a"`, `"b"`, etc. in the example above.  However, `cPtrToRubyString` has a
significant issue: the code near the end assumes that you are currently
emitting LLVM IR into the current Ruby function being compiled.  So when you
are constructing the LLVM IR for `buildGlobalHashLiteral`, above, you hit
the `ENFORCE` in `cPtrToRubyString` when emitting code to reference the
`VALUE`s for the constant strings in the hash table.

That's the problem.  Why is the solution of removing the `ENFORCE`s
reasonable?

It looks as though the `ENFORCE`s were there because the assumption is that
if you are going to reference string constants (resp. symbols via
`Payload::idIntern`), we might as well just grab references to them at the
top of the current Ruby function.  Grabbing them at the entrance to the
function relies on the compiler for two things: de-duplication of multiple
references to the same string constant and code-sinking the references if
the reference only occurs on some paths through the function.

But if we're going to rely on the compiler for such things, we might as well
grab the references directly at the use sites and rely on the compiler to
(again) eliminate multiple references where appropriate and to perform the
appropriate common subexpression elimination and hoist the references where
it is beneficial (e.g. out of loops or similar).  There doesn't seem to be
any benefit to ensuring that we have references at the start of the current
function, and it is actively harmful to use cases that want to treat strings
or symbols identically to other literals such as integers or floats.

We're not performing the same removal in `cPtrToRubyRegexp` at the present
time because there doesn't seem to be a need, but it would not be entirely
surprising if a similar use case to the above came up.

The alternative to removing these `ENFORCE`s is to split up the construction
of the LLVM IR instructions from the inserting of the LLVM IR instructions
into the relevant functions.  This path is doable, but having tried it, it
quickly gets messy and the proposed path here is much cleaner.

* update test expectations