Typecase function that uses tags on tagged pointer (#3605)
Adds a `typecase()` function that uses tags on tagged pointer. Now we can use `typecase()` with tagged pointers, which uses `tag()` to determine if a cast is legal!

I have made the specialization work with both `TypePtr` and `TreePtr`. To make the same code work with both types, those types had to implement the following interface:

* `isa<T>` static template function. These forward to `isa_type`/`isa_tree`, respectively, with some specializations to support casting to `TypePtr`/`TreePtr`/`Expression`.
* `cast_nonnull<T>` static template function. These forward to `cast_type_nonnull`/`cast_tree_nonnull`, respectively, with some specializations.

One change to the interface is that lambdas are called with a _reference type_, which I think is cleaner anyway. Hence the big diff with shallow changes.

The tagged `typecase` version takes a reference type rather than a pointer type, letting the existing (non-tagged) version coexist peacefully with it. I added in a friendly `static_assert` that checks for `.tag()` and complains if you use the wrong version. No one will be able to use the non-tagged typecase on `TypePtr`/`TreePtr`.

I've verified via godbolt that this typecase implementation optimizes into a simple jump table: https://godbolt.org/z/76MEhM