Method rename support (initial implementation)
Implements a first cut of method rename support; builds on the
constant rename work. This implementation supports subclasses but
doesn't handle union types, mixins, and probably more cases.

The goal of this change is to enable method renames in experimental
mode and gather some user feedback about which missing bits are most
important.