finish splitting up the payload (#231)
* put a separate copy of sorbet_ensure_arity in the code payload

We're going to need this for argument checking in both places.

* move plus/minus fastpaths to the code payload

* move lt/gt fast paths to the payload

This change is not quite as straightforward as one might like, because gt
didn't really *have* a fastpath.  So we invented one.  That also means that
we have to send ge/le through the "slow" path, which isn't really slow, it's
just practically guaranteed to not get inlined.  This arrangement means that
ge/le are probably a performance trap in a large number of cases, because
the fixnum/fixnum case won't get inlined at all.  We should fix this.

* move sorbet_callBlock to the code payload

* move local variable read/write to the code payload

The only complication here is renaming `vm_env_write_slowpath`, because we
are going to need it to be accessible in `libruby.so`, but we can't call it
`vm_env_write_slowpath` due to name conflicts.  So we tack on the `sorbet_`
prefix.

* move array operations to the code payload

* move operations on hashes and simple name-based intrinsics to the code payload

`sorbet_buildHashIntrinsic` needed to come along to be able to call
`sorbet_hashBuild`, and at that point, `sorbet_build{Array,Range}Intrinsic`
both looked simple enough that they could stand to be inlined.

* move inline call helpers to the code payload

* delete unused sorbet_rubyException

* make sorbet_rubyUndef local to the payload for better hygiene

...but `sorbet_rubyUndef` also needs to be in the code payload, because the
compiler uses it to construct undefined values.  So make a copy in the code
payload also.

* move sorbet_rubyTopSelf to the code payload

...and inline it at its lone use in the VM payload.

* move sorbet_setRubyStackFrame to the code payload

* move sorbet_callSuper to the code payload

* close the performance hole for numeric le/ge

We would never inline numeric le/ge with the current setup; now we do, for
the fixnum/fixnum case.

* don't compile the payload for the code payload anymore

* sorbet_inlineCacheInvalidated has to be a public function now

* make sure that all our external functions are kept alive in bitcode

* add a couple liveness definitions for rb_* functions

* move various functions back to the code payload

These functions were getting inlined before according to .exp files; let's
keep them inlined.

* move more predicates into the code payload

* remove no-longer needed SORBET_LLVM_PAYLOAD define

* adjust payload's view of the world

We're no longer compiling this file with `SORBET_LLVM_PAYLOAD`, and so we
need to adjust various bits of code accordingly.

* don't declare sorbet_stackoverflow as always inline

* move bool functions to the code payload

* move more integer intrinsics and sorbet_boolToRuby over

* move array intrinsics, T.{unsafe,must} over

* move hash intrinsics over

* move sorbet_{bang,selfNew} over to the code payload

* sorbet_{isa,isa_class_of} already live in the code payload

* move kw/rest arg handling functions to the code payload

* move sorbet_getMethodBlockAsProc to the code payload

* move global variable get/set to the code payload

* move some exception-handling bits to the code payload

* move sorbet_checkStack to the code payload

* move sorbet_isa_{Proc,RootSingleton} to the code payload

* rb_errinfo and rb_set_errinfo are held alive by more conventional means now

* update test expectations in one go

* format the code payload once again