Forbid making final method public in child
This would already raise:

```ruby
class Parent
  sig(:final) {void}
  def foo; end
end

class Child < Parent
  private :foo
end
```

(public in parent made private in child)

But if that were flipped around (private in parent, public in child),
for some reason that was allowed.

I don't think this should be asymmetrical. If something is final, it
should not be modifiable.