Show all method flags in symbol-table output (#3681)
* Add some tests to showcase it better

* Show all method flags in symbol-table output

I chose a representation that wouldn't cause a ton of churn. Namely: if
you're a method symbol and you're public, we don't print that you're
public (this is what was already done). Also I chose to use the `|`
separator between flags because it reminds me of bitflags, which these
flags actually are under the hood.