Dispatch <Magic>.<self-new> through new, not initialize (#3682)
* Add test of current behavior

* Dispatch through new in <Magic>.<self-new>, not initialize

* Fix failing LSP test

Now that we included Class#new in the dispatch result, a hover result
changed from `(nothing)` to something, but it looked bad, because
`Class#new` is entered as a "synthetic symbol" (i.e., any method entered
in GlobalState.cc).

That heuristic was too broad for what this code wanted to do, which is
show `{foo: Integer(0)}` when you hover on a hash instead of
`<Magic>.<build-hash>`, etc. (there are a lot `<Magic>.<build-*>`).

So instead, I check whether the owner is `T.class_of(<Magic>)` because
that's a better indication that the thing shouldn't be printed.