Introduce ConstSymbolData for const references to Symbols.
`const SymbolData` is trivially convertable to a `SymbolData`, which permits mutations to `Symbol`s even with a `const GlobalState&`!

This change locks that down so that this escape hatch is not viable.

It includes an explicit `const_cast` to permit code that relied on this behavior until it can be fully fixed.