Method renames: robustness in send expression handling (#3685)
When renaming a method we have to update send expressions; unfortunately a send expression as returned in a SendResponse contains the entire <receiver>.<method><args> expression. Before this change, we were parsing the receiver expression to skip over it to find the method; this was brittle and broken in a many cases.

In this change we add receiver location info to the SendResponse query response type, and use that to find the method name in a send expression more reliably. This removes all instances of search-and-replace from method renames.