Adds support for prop/const/attr_* in Find All References / Document Highlight (#3739)
* Support prop/const/attr_* in Find All References and Document Highlight.

Adds support for prop/const/attr_* in Find All References / Document Highlight. Involves rejiggering document highlight's helper functions a bit so they can cleanly re-use deduping functionality. Also fixes a bug in deduping where `auto` inferred an unsigned type when it should've been an `int`.

When you Find All References on a field reference, method call, or a method definition site, you could plausibly be pointing at a method or a field defined by a `prop` or an `attr_reader`. The code does the following to detect if this is true:
* Extracts the base name of the thing you're pointing at (e.g. `@foo` => `foo`, `foo=` => `foo`)
* Ensures one of method `foo` or `foo=` exists on the owning class.
* Checks the definition site of `foo` or `foo=`.
  * If it begins with `prop`/etc, we're in business.
* Returns a struct containing the type of accessor (reader / writer / both / none) and symbols for the field, reader, and writer.

I've added code that consumes this struct and returns all references that refer to any of the items defined by `prop` et al.