Method renames: handle secondary dispatch components (union types) (#3757)
While searching for related symbols in method renames, handle multiple dispatch results in sends. When we find a secondary dispatch result, we follow the dispatch result linked list, adding any new methods we find. You can think of this code as the traversal of a graph where nodes are methods, and edges exist between methods that are related either by overriding or by being used in a union type.