Make ResolveSignaturesWalk a ShallowMap and small code tweaks. (#3771)
Make ResolveSignaturesWalk a ShallowMap and small code tweaks. The send nodes that ResolveSignaturesWalk was processing are readily available earlier in the type params pass.

Addresses lingering feedback in #3762 from @jez and @elliottt.

Should result in a speedup as ResolveSignaturesWalk is a serial pass over all of the ASTs, so it doing less work is significant on the critical path.