Break up name table by name type. (#3787)
The name table currently contains a vector of 24-byte Name objects. A full 8 bytes of this object is devoted to a one-byte tag, which indicates if the name is a UTF8Name, ConstantName, or a UniqueName. Furthermore, ConstantName and UniqueNames take fewer than 16 bytes and have even more padding.

This PR splits the name table into separate vectors for UTF8Names, ConstantNames, and UniqueNames. UTF8Names are 16 bytes (33% savings), ConstantNames are 4 bytes (83% savings), and UniqueNames are 12 bytes (50% savings).

On large codebases typechecked on machines with many cores and with a cached name table, this PR reduces max RSS induced during the parallel indexing phase, which duplicates the name table once per core. This change also causes a small typechecking speedup.