Some <S <S ...$1> $1> classes are gone nowo
These classes are singleton classes of the singleton class.
The appear to have been materialized as the result of adding a modifier
for a method that doesn't exist on the singleton class of the singleton
class.

The namer logic forces that class into existence, and but finds no
method there to make private (because it doesn't exist there).

I argue that this diff is actually more correct than it was before,
because this indicates to me that in a world where #3666 has landed,
this would have tried to register a ZSuper method for `:bar` on the
A.singleton_class.singleton_class, which is "more" wrong than the
decision to not look at `private_class_method` calls inside methods.