Empty hashes are containers of untyped
Largely, this is to preserve the existing behavior.

Now, this untypedness is handled consistently for both tuples and shapes
in `underlying`.

This strikes a balance--for places that actually try to look at whether
you have a shape type or not, it's still visibly an empty shape type
(e.g., some keyword arg stuff was depending on this). But if you ever
try to call a method on it that isn't implemented as a
ShapeType/TupleType intrinsic (and thus falls back up to Hash or Array),
then it will decay as an untyped container.