Hide origins2Explanations for <self>
Giving the `Cast` instruction a value of `md.loc` gave the weird
impression that the "origin" of `<self>` was either:

- the whole method body
- the whole file contents (for self at the top level)
- the containing ClassDef declLoc

depending on the situation. In almost all of these, the error is going
to be useless:

- showing a whole method body is just going to truncate the method with
  an elipsis for most real-world methods
- same for the whole file contents--makes no sense
- the containing ClassDef declLoc is _maybe_ something the user could
  make sense of, but probably we want to say something more explicit,
  like "we got this type because it was `self`"

In particular, none of those locs would be able to participate in
anything resembling an autocorrect to an assignment in the source code.

Thus, I've removed it.